added further menu validations

+++ lib/app / models/ride.rb
@@ -26,21 +26,31 @@ end
 def self.see_leaderboard
     puts ""
     puts ""
-    puts "Alright, so your time was #{self.last.time} seconds on your last run on #{self.last.trail.name}. Here's the leaderboard in seconds for that trail, that has a difficulty level of #{Ride.last.trail.difficulty}:"
+    unless Shredder.last.name == Ride.last.shredder.name
+        puts "[If you had any rides, you'd be able to see the trail leaderboard for your last run here...]"
+        puts ""
+    else
+        puts "Alright, so your time was #{self.last.time} seconds on your last run on #{self.last.trail.name}. Here's the leaderboard in seconds for that trail, that has a difficulty level of #{Ride.last.trail.difficulty}:"
+        puts ""
+        sleep 3
+        rides = self.where(trail: self.last.trail)
+        puts rides.map {|ride| [ride.time, ride.shredder.name]}.sort
+        sleep 4
+    end
     puts ""
-    sleep 3
-    rides = self.where(trail: self.last.trail)
-    puts rides.map {|ride| [ride.time, ride.shredder.name]}.sort
-    puts ""
-    sleep 5
+    sleep 1
 end
 
 # delete_ride
 def self.delete_ride
     puts ""
     puts ""
-    puts "Yeah, we didn't want to put you on blast but that last run on #{self.last.trail.name} was pretty weak. We're deleting it from the mountain record book...slowpoke."
-    self.last.destroy
+    unless Shredder.last.name == Ride.last.shredder.name
+            puts "[If you had any rides, you'd be able to delete your last one here...Get some rides in yo!]"
+    else
+            puts "Yeah, we didn't want to put you on blast but that last run on #{self.last.trail.name} was pretty weak. We're deleting it from the mountain record book...slowpoke."
+            self.last.destroy
+    end
     puts ""
     puts ""
     sleep 3
